Added UniformBuffer CalculateAlignedSize function 09/22/2021 | 21w38a1 | 0.7.86

+++ TRAP/src/Graphics/API/RendererAPI.h
@@ -1291,10 +1291,10 @@ namespace TRAP::Graphics
 
 		inline static struct GPUSettings
 		{
-			uint32_t UniformBufferAlignment;
+			uint64_t UniformBufferAlignment;
 			uint32_t MaxUniformBufferRange;
-			uint32_t UploadBufferTextureAlignment;
-			uint32_t UploadBufferTextureRowAlignment;
+			uint64_t UploadBufferTextureAlignment;
+			uint64_t UploadBufferTextureRowAlignment;
 			uint32_t MaxVertexInputBindings;
 			uint32_t MaxVertexInputAttributes;
 			uint32_t MaxRootSignatureDWORDS;
