+++ TRAP/src/Graphics/API/ResourceLoader.h
@@ -65,7 +65,7 @@ namespace TRAP::Graphics::API
 
 		static RendererAPI::ResourceLoaderDesc DefaultResourceLoaderDesc;
 		static void StreamerThreadFunc(ResourceLoader* loader);
-		static uint32_t UtilGetTextureRowAlignment();
+		static uint64_t UtilGetTextureRowAlignment();
 		static uint32_t UtilGetTextureSubresourceAlignment(TRAP::Graphics::API::ImageFormat fmt = TRAP::Graphics::API::ImageFormat::Undefined);
 		static uint32_t UtilGetSurfaceSize(TRAP::Graphics::API::ImageFormat fmt, uint32_t width, uint32_t height,
 		                                   uint32_t depth, uint32_t rowStride, uint32_t sliceStride,
